+++ samha.backend/src/main/java/com/samha/domain/RestricaoProfessor.java
@@ -19,7 +19,7 @@ import javax.persistence.Table;
 public class RestricaoProfessor extends BaseLogEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     
     @Column(nullable = false)
     private String nome;
@@ -58,12 +58,11 @@ public class RestricaoProfessor extends BaseLogEntity{
     @JoinColumn(name = "professor_id", nullable = false)
     private Professor professor;
 
-
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
