+++ samha.backend/src/main/java/com/samha/persistence/IAlocacaoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IAlocacaoRepository extends JpaRepository<Alocacao, Integer> {
+public interface IAlocacaoRepository extends JpaRepository<Alocacao, Long> {
 }
