+++ samha.backend/src/main/java/com/samha/persistence/IProfessorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IProfessorRepository extends JpaRepository<Professor, Integer> {
+public interface IProfessorRepository extends JpaRepository<Professor, Long> {
 }
